requirements.txt
Django==4.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-twitter==3.5
facebook-sdk==3.1.0
django-allauth==0.58.2
chart.js==2.9.4